Remove `mutable` field from action types
Similar in spirit to #58938.

The app router reducer state used to be managed by useReducer, so it was
written to be resilient to rebasing — the same action may be processed
multiple times. Now that we've lifted the reducer outside of React
(#56497), each action runs only a single time. So we can simplify some
of the logic.

The purpose of the `mutable` field was so that if an action is
processed multiple times, state could be reused between each run; for
example, to prevent redundant network fetches. Now that this scenario
can no longer happen, we can remove it.

I had to update some of the unit tests in navigate-reducer because they
were written with the assumption that the reducer was called
multiple times. As far as I can tell, most of this behavior is covered
by e2e tests anyway, so I don't think it's too risky.