Make CacheNodeSeedData match FlightRouterState more closely
When PPR is enabled, the CacheNodeSeedData and FlightRouterState
contained in an RSC payload (whether it's static or dynamic) should
always have the same tree structure.

The only reason the two trees would ever be different is if the
server partially renders with `loading.tsx`, but that behavior doesn't
exist when PPR is enabled. (In that case, the CacheNodeSeedData would
describe only a subset of FlightRouterState.)

The behavior is already correct but to make the types a bit closer,
this changes the `parallelRoutes` slot of CacheNodeSeedData to be
non-nullable; if there are no children, it should be an empty object.

It's not semantically important, I'm only changing it to make the types
more consistent, since we typically traverse both trees in parallel.

Eventually we will probably merge these into a single transport type.