Add prefetchHead to Cache Node type
Like segment data, the head data for a segment may have two versions:
a static version with dynamic holes, and a full version with everything.

This adds a `prefetchHead` field to Cache Node. It works the same as the
`prefetchRsc` field.

As with the previous commit, we use `useDeferredValue` to determine
which version to render.

One awkward bit is that although the head element belongs to a
particular segment, it's always rendered at the top of the App, by App
Router. I believe this is to allow the page metadata to update even
before the page itself appears. (Though I wonder if we could render this
in the shared layout instead.) In the meantime, in order for
useDeferredValue to work, we need to remount the head element whenever
the segment changes.