[PPR Navs] Remove fallback to applyFlightData
The PPR version of navigateReducer falls back to the old implementation
for processing server payloads, applyFlightData, in some cases, like
when the prefetch response is stale.

As an incremental step to start migrating this over, and to get more
dogfooding coverage, this PR removes that fallback so that the code
always goes through the PPR implementation.

Because of how the code is factored, for now this means we will always
issue a static prefetch for a navigation, even if it's not already
available before the navigation happens. Longer term I don't think this
is necessarily the right trade off — if there's no prefetched response,
it's probably better to do what we do for initial render and have the
server stitch together the dynamic and static segments together into a
single response. But I think this makes sense an incremental step and
will allow me to keep incrementally moving stuff to the
new implementation.