Rewrite run-on sentence for clarity; remove 'default' from wording
Broke up a run-on sentence that's difficult to understand. Furthermore, the skipTrailingSlashRedirect as I understand it does not only disable "default redirects" as was stated. It disables slash redirecting altogether, whether the default behavior of removing them or the opposite behavior of keeping them when set by the trailingSlash flag. Thus, removed the word 'default'.